+++ designer/client/src/actions/nk/nodeDetails.ts
@@ -52,32 +52,30 @@ export function nodeDetailsClosed(nodeId: string): NodeDetailsClosed {
 //we don't return ThunkAction here as it would not work correctly with debounce
 //TODO: use sth better, how long should be timeout?
 const validate = debounce(
-    async (processName: string, validationRequestData: ValidationRequest, callback: (data: ValidationData, nodeId: NodeId) => void) => {
+    async (
+        processName: string,
+        validationRequestData: ValidationRequest,
+        callback: (nodeId: NodeId, data?: ValidationData | void) => void,
+    ) => {
+        const validate = (node: NodeType) =>
+            NodeUtils.nodeIsProperties(node)
+                ? //NOTE: we don't validationRequestData contains processProperties, but they are refreshed only on modal open
+                  HttpService.validateProperties(processName, { additionalFields: node.additionalFields, name: node.id })
+                : HttpService.validateNode(processName, { ...validationRequestData, nodeData: node });
+
         const nodeId = validationRequestData.nodeData.id;
         const nodeWithChangedName = applyIdFromFakeName(validationRequestData.nodeData);
-        if (NodeUtils.nodeIsProperties(nodeWithChangedName)) {
-            //NOTE: we don't validationRequestData contains processProperties, but they are refreshed only on modal open
-            const { data } = await HttpService.validateProperties(processName, {
-                additionalFields: nodeWithChangedName.additionalFields,
-                name: nodeWithChangedName.id,
-            });
-            callback(data, nodeId);
-        } else {
-            const { data } = await HttpService.validateNode(processName, {
-                ...validationRequestData,
-                nodeData: nodeWithChangedName,
-            });
-            callback(data, nodeId);
-        }
+        const data = await validate(nodeWithChangedName);
+        callback(nodeId, data);
     },
     500,
 );
 
 export function validateNodeData(processName: string, validationRequestData: ValidationRequest): ThunkAction {
     return (dispatch, getState) => {
-        validate(processName, validationRequestData, (data, nodeId) => {
+        validate(processName, validationRequestData, (nodeId, data) => {
             // node details view creates this on open and removes after close
-            if (getState().nodeDetails[nodeId]) {
+            if (data && getState().nodeDetails[nodeId]) {
                 dispatch(nodeValidationDataUpdated(nodeId, data));
             }
         });