+++ designer/deployment-manager-api/src/main/scala/pl/touk/nussknacker/engine/api/deployment/cache/CachingProcessStateDeploymentManager.scala
@@ -24,10 +24,12 @@ class CachingProcessStateDeploymentManager(delegate: DeploymentManager, cacheTTL
     .expireAfterWrite(java.time.Duration.ofMillis(cacheTTL.toMillis))
     .buildAsync[ProcessName, List[StatusDetails]]
 
-  override def getProcessState(idWithName: ProcessIdWithName, lastStateAction: Option[ProcessAction])(
-      implicit freshnessPolicy: DataFreshnessPolicy
-  ): Future[WithDataFreshnessStatus[ProcessState]] =
-    delegate.getProcessState(idWithName, lastStateAction)
+  override def resolve(
+      idWithName: ProcessIdWithName,
+      statusDetails: List[StatusDetails],
+      lastStateAction: Option[ProcessAction]
+  ): Future[ProcessState] =
+    delegate.resolve(idWithName, statusDetails, lastStateAction)
 
   override def getProcessStates(
       name: ProcessName
@@ -123,18 +125,7 @@ object CachingProcessStateDeploymentManager extends LazyLogging {
           )
         )
       logger.debug(s"Wrapping DeploymentManager: $delegate with caching mechanism with TTL: $cacheTTL")
-      delegate match {
-        case postprocessing: PostprocessingProcessStatus =>
-          new CachingProcessStateDeploymentManager(delegate, cacheTTL) with PostprocessingProcessStatus {
-            override def postprocess(
-                idWithName: ProcessIdWithName,
-                statusDetailsList: List[StatusDetails]
-            ): Future[Option[ProcessAction]] =
-              postprocessing.postprocess(idWithName, statusDetailsList)
-          }
-        case _ =>
-          new CachingProcessStateDeploymentManager(delegate, cacheTTL)
-      }
+      new CachingProcessStateDeploymentManager(delegate, cacheTTL)
     } else {
       logger.debug(s"Skipping ProcessState caching for DeploymentManager: $delegate")
       delegate